feat #1: use socket.io
