+++ src/classes/Player.js
@@ -1,4 +1,4 @@
-class Player {
+export default class Player {
   constructor(x, y) {
     this.x = x;
     this.y = y;
